src/pyproject.toml
[tool.poetry]
name = "djlodging"
version = "0.1.0"
description = "Django backend for the lodging reservation application"
authors = ["Kostiantyn Perun <kosperun@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "4.0"
django-environ = "^0.8.1"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.0.0"
drf-spectacular = "^0.23.1"
Pillow = "^9.0.0"
psycopg2-binary = "^2.9.3"
django-health-check = "^3.16.5"
sendgrid = "^6.9.7"
stripe = "^4.1.0"
dj-stripe = "^2.6.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pylint-django = "^2.5.3"
django-silk = "^5.0.1"
factory-boy = "^3.2.1"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
Faker = "^14.0.0"
ipython = "^8.4.0"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--reuse-db"
testpaths = "tests"
DJANGO_SETTINGS_MODULE = "djlodging.django_app.django_core.settings"

[tool.poetry.scripts]
manage = 'djlodging.django_app.manage:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
